Tumor rate and spectral data

This dataset feature tumor rate data with spectral data descriptors. It is used as an example in the second vignette of the package.

colon

Format

A data frame with 80 observations on the following 180 variables.

X..Tumor.Cells

a numeric vector

X4.69499969

a numeric vector

X4.68499947

a numeric vector

X4.67499971

a numeric vector

X4.66499949

a numeric vector

X4.65499973

a numeric vector

X4.6449995

a numeric vector

X4.63499975

a numeric vector

X4.62499952

a numeric vector

X4.61499977

a numeric vector

X4.60499954

a numeric vector

X4.59499979

a numeric vector

X4.58499956

a numeric vector

X4.57499981

a numeric vector

X4.56499958

a numeric vector

X4.55499983

a numeric vector

X4.5449996

a numeric vector

X4.53499985

a numeric vector

X4.52499962

a numeric vector

X4.51499987

a numeric vector

X4.50499964

a numeric vector

X4.49499989

a numeric vector

X4.48499966

a numeric vector

X4.4749999

a numeric vector

X4.46499968

a numeric vector

X4.45499992

a numeric vector

X4.44499969

a numeric vector

X4.43499947

a numeric vector

X4.42499971

a numeric vector

X4.41499949

a numeric vector

X4.40499973

a numeric vector

X4.3949995

a numeric vector

X4.38499975

a numeric vector

X4.37499952

a numeric vector

X4.36499977

a numeric vector

X4.35499954

a numeric vector

X4.34499979

a numeric vector

X4.33499956

a numeric vector

X4.32499981

a numeric vector

X4.31499958

a numeric vector

X4.30499983

a numeric vector

X4.2949996

a numeric vector

X4.28499985

a numeric vector

X4.27499962

a numeric vector

X4.26499987

a numeric vector

X4.25499964

a numeric vector

X4.24499989

a numeric vector

X4.23499966

a numeric vector

X4.2249999

a numeric vector

X4.21499968

a numeric vector

X4.20499992

a numeric vector

X4.19499969

a numeric vector

X4.18499994

a numeric vector

X4.17499971

a numeric vector

X4.16499949

a numeric vector

X4.15499973

a numeric vector

X4.1449995

a numeric vector

X4.13499975

a numeric vector

X4.12499952

a numeric vector

X4.11499977

a numeric vector

X4.10499954

a numeric vector

X4.09499979

a numeric vector

X4.08499956

a numeric vector

X4.07499981

a numeric vector

X4.06499958

a numeric vector

X4.05499983

a numeric vector

X4.0449996

a numeric vector

X4.03499985

a numeric vector

X4.02499962

a numeric vector

X4.01499987

a numeric vector

X4.00499964

a numeric vector

X3.99499965

a numeric vector

X3.98499966

a numeric vector

X3.97499967

a numeric vector

X3.96499968

a numeric vector

X3.95499969

a numeric vector

X3.94499969

a numeric vector

X3.9349997

a numeric vector

X3.92499971

a numeric vector

X3.91499972

a numeric vector

X3.90499973

a numeric vector

X3.89499974

a numeric vector

X3.88499975

a numeric vector

X3.87499976

a numeric vector

X3.86499977

a numeric vector

X3.85499978

a numeric vector

X3.84499979

a numeric vector

X3.8349998

a numeric vector

X3.82499981

a numeric vector

X3.81499982

a numeric vector

X3.80499983

a numeric vector

X3.7949996

a numeric vector

X3.78499961

a numeric vector

X3.77499962

a numeric vector

X3.76499963

a numeric vector

X3.75499964

a numeric vector

X3.74499965

a numeric vector

X3.73499966

a numeric vector

X3.72499967

a numeric vector

X3.71499968

a numeric vector

X3.70499969

a numeric vector

X3.69499969

a numeric vector

X3.6849997

a numeric vector

X3.67499971

a numeric vector

X3.66499972

a numeric vector

X3.65499973

a numeric vector

X3.64499974

a numeric vector

X3.63499975

a numeric vector

X3.62499976

a numeric vector

X3.61499977

a numeric vector

X3.60499978

a numeric vector

X3.59499979

a numeric vector

X3.5849998

a numeric vector

X3.57499981

a numeric vector

X3.56499982

a numeric vector

X3.55499983

a numeric vector

X3.54499984

a numeric vector

X3.53499961

a numeric vector

X3.52499962

a numeric vector

X3.51499963

a numeric vector

X3.50499964

a numeric vector

X3.49499965

a numeric vector

X3.48499966

a numeric vector

X3.47499967

a numeric vector

X3.46499968

a numeric vector

X3.45499969

a numeric vector

X3.44499969

a numeric vector

X3.4349997

a numeric vector

X3.42499971

a numeric vector

X3.41499972

a numeric vector

X3.40499973

a numeric vector

X3.39499974

a numeric vector

X3.38499975

a numeric vector

X3.37499976

a numeric vector

X3.36499977

a numeric vector

X3.35499978

a numeric vector

X3.34499979

a numeric vector

X3.3349998

a numeric vector

X3.32499981

a numeric vector

X3.31499982

a numeric vector

X3.30499983

a numeric vector

X3.29499984

a numeric vector

X3.28499961

a numeric vector

X3.27499962

a numeric vector

X3.26499963

a numeric vector

X3.25499964

a numeric vector

X3.24499965

a numeric vector

X3.23499966

a numeric vector

X3.22499967

a numeric vector

X3.21499968

a numeric vector

X3.20499969

a numeric vector

X3.19499969

a numeric vector

X3.1849997

a numeric vector

X3.17499971

a numeric vector

X3.16499972

a numeric vector

X3.15499973

a numeric vector

X3.14499974

a numeric vector

X3.13499975

a numeric vector

X3.12499976

a numeric vector

X3.11499977

a numeric vector

X3.10499978

a numeric vector

X3.09499979

a numeric vector

X3.0849998

a numeric vector

X3.07499981

a numeric vector

X3.06499982

a numeric vector

X3.05499983

a numeric vector

X3.04499984

a numeric vector

X3.03499985

a numeric vector

X3.02499962

a numeric vector

X3.01499963

a numeric vector

X3.00499964

a numeric vector

X2.99499965

a numeric vector

X2.98499966

a numeric vector

X2.97499967

a numeric vector

X2.96499968

a numeric vector

X2.95499969

a numeric vector

X2.94499969

a numeric vector

X2.9349997

a numeric vector

X2.92499971

a numeric vector

X2.91499972

a numeric vector

References

Régression Bêta PLS. (French) [PLS Beta regression.], F. Bertrand, N. Meyer, M. Beau-Faller, K. El Bayed, N. Izzie-J., M. Maumy-Bertrand, (2013), J. SFdS, 154(3):143-159

Partial Least Squares Regression for Beta Regression Models. F. Bertrand, M. Maumy (2021). useR! 2021, Zurich.

Examples

data(colon)
str(colon)

#> 'data.frame':	80 obs. of  180 variables:
#>  $ X..Tumor.Cells: num  0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 ...
#>  $ X4.69499969   : num  8.99e-05 -8.83e-05 -1.38e-05 7.74e-04 -1.48e-05 ...
#>  $ X4.68499947   : num  8.54e-05 -9.43e-05 9.67e-06 7.21e-04 -1.08e-04 ...
#>  $ X4.67499971   : num  1.50e-04 -8.04e-05 8.10e-05 8.56e-04 -1.12e-04 ...
#>  $ X4.66499949   : num  3.59e-04 -3.75e-05 2.44e-04 1.30e-03 1.60e-04 ...
#>  $ X4.65499973   : num  0.000858 0.00017 0.00061 0.00421 0.002897 ...
#>  $ X4.6449995    : num  3.87e-04 3.59e-05 2.61e-04 3.05e-03 1.55e-03 ...
#>  $ X4.63499975   : num  0.000596 0.000147 0.000641 0.002869 0.001939 ...
#>  $ X4.62499952   : num  2.19e-04 -1.03e-05 1.45e-04 8.88e-04 2.46e-04 ...
#>  $ X4.61499977   : num  2.30e-04 -2.13e-05 1.97e-04 7.91e-04 1.04e-04 ...
#>  $ X4.60499954   : num  2.72e-04 1.92e-05 4.53e-04 8.31e-04 2.37e-04 ...
#>  $ X4.59499979   : num  4.90e-04 9.56e-05 9.26e-04 1.30e-03 8.26e-04 ...
#>  $ X4.58499956   : num  0.00092 0.000286 0.001638 0.001838 0.001668 ...
#>  $ X4.57499981   : num  0.001012 0.000348 0.001851 0.002019 0.001903 ...
#>  $ X4.56499958   : num  0.000713 0.000218 0.001074 0.001364 0.001047 ...
#>  $ X4.55499983   : num  3.83e-04 4.46e-05 3.68e-04 7.45e-04 2.24e-04 ...
#>  $ X4.5449996    : num  2.79e-04 -2.30e-05 2.23e-04 5.12e-04 5.82e-05 ...
#>  $ X4.53499985   : num  2.15e-04 -3.79e-05 2.37e-04 4.67e-04 -8.30e-07 ...
#>  $ X4.52499962   : num  0.000966 0.000245 0.001627 0.00247 0.001954 ...
#>  $ X4.51499987   : num  0.001216 0.000362 0.001684 0.001861 0.001258 ...
#>  $ X4.50499964   : num  3.02e-04 2.53e-05 3.19e-04 7.99e-04 1.61e-04 ...
#>  $ X4.49499989   : num  3.15e-04 3.95e-06 2.30e-04 6.48e-04 1.98e-05 ...
#>  $ X4.48499966   : num  3.01e-04 1.65e-05 3.41e-04 5.78e-04 1.67e-04 ...
#>  $ X4.4749999    : num  3.67e-04 2.14e-05 2.67e-04 6.50e-04 1.92e-04 ...
#>  $ X4.46499968   : num  4.10e-04 1.26e-05 2.87e-04 6.40e-04 7.65e-05 ...
#>  $ X4.45499992   : num  3.75e-04 3.56e-05 4.47e-04 7.15e-04 1.92e-04 ...
#>  $ X4.44499969   : num  5.34e-04 8.22e-05 5.52e-04 8.56e-04 4.96e-04 ...
#>  $ X4.43499947   : num  0.000661 0.000124 0.000612 0.000984 0.000617 ...
#>  $ X4.42499971   : num  0.000582 0.00013 0.000809 0.001099 0.000516 ...
#>  $ X4.41499949   : num  0.000554 0.000158 0.000919 0.001172 0.000603 ...
#>  $ X4.40499973   : num  0.000527 0.000138 0.000954 0.001048 0.000667 ...
#>  $ X4.3949995    : num  0.000559 0.000119 0.000974 0.000949 0.000563 ...
#>  $ X4.38499975   : num  0.000615 0.000165 0.001093 0.001322 0.000648 ...
#>  $ X4.37499952   : num  0.000751 0.000234 0.001365 0.00197 0.000956 ...
#>  $ X4.36499977   : num  0.000768 0.000234 0.001271 0.001762 0.000851 ...
#>  $ X4.35499954   : num  0.000808 0.000213 0.001216 0.001665 0.000708 ...
#>  $ X4.34499979   : num  0.001221 0.000331 0.001931 0.002355 0.001624 ...
#>  $ X4.33499956   : num  0.001829 0.000502 0.003084 0.003697 0.003243 ...
#>  $ X4.32499981   : num  0.002211 0.000722 0.004042 0.004747 0.004498 ...
#>  $ X4.31499958   : num  0.001843 0.000609 0.0034 0.003802 0.003518 ...
#>  $ X4.30499983   : num  0.001308 0.000529 0.002397 0.002966 0.00229 ...
#>  $ X4.2949996    : num  0.00101 0.00039 0.00173 0.00208 0.00141 ...
#>  $ X4.28499985   : num  0.0011 0.00039 0.00185 0.00213 0.00143 ...
#>  $ X4.27499962   : num  0.001184 0.000436 0.002059 0.002134 0.001581 ...
#>  $ X4.26499987   : num  0.001126 0.000454 0.002035 0.002173 0.001327 ...
#>  $ X4.25499964   : num  0.001059 0.000412 0.001895 0.001904 0.001288 ...
#>  $ X4.24499989   : num  0.000899 0.000283 0.001539 0.001531 0.000918 ...
#>  $ X4.23499966   : num  0.0008 0.000187 0.001201 0.001265 0.000575 ...
#>  $ X4.2249999    : num  0.000762 0.000131 0.001096 0.001066 0.000399 ...
#>  $ X4.21499968   : num  0.000711 0.000108 0.000945 0.00112 0.000334 ...
#>  $ X4.20499992   : num  0.000693 0.000111 0.000945 0.001135 0.000223 ...
#>  $ X4.19499969   : num  0.000804 0.000172 0.00164 0.001528 0.000459 ...
#>  $ X4.18499994   : num  0.001246 0.000522 0.003401 0.002819 0.001436 ...
#>  $ X4.17499971   : num  0.00214 0.00108 0.00545 0.0044 0.00308 ...
#>  $ X4.16499949   : num  0.00279 0.00138 0.00529 0.00655 0.00422 ...
#>  $ X4.15499973   : num  0.00251 0.00101 0.00365 0.00668 0.00348 ...
#>  $ X4.1449995    : num  0.00371 0.00147 0.00797 0.00822 0.00499 ...
#>  $ X4.13499975   : num  0.00573 0.00211 0.01354 0.01176 0.00862 ...
#>  $ X4.12499952   : num  0.00593 0.00268 0.01487 0.01133 0.00941 ...
#>  $ X4.11499977   : num  0.00808 0.0038 0.01942 0.01456 0.01432 ...
#>  $ X4.10499954   : num  0.00405 0.00158 0.00821 0.00709 0.00651 ...
#>  $ X4.09499979   : num  0.002255 0.000585 0.003045 0.002778 0.002261 ...
#>  $ X4.08499956   : num  0.002157 0.000229 0.001354 0.001741 0.000784 ...
#>  $ X4.07499981   : num  0.003341 0.000475 0.00207 0.002829 0.001476 ...
#>  $ X4.06499958   : num  0.00594 0.00129 0.00587 0.00846 0.00733 ...
#>  $ X4.05499983   : num  0.01216 0.00299 0.01732 0.01715 0.01925 ...
#>  $ X4.0449996    : num  0.00717 0.00108 0.00794 0.00597 0.0066 ...
#>  $ X4.03499985   : num  0.00417 0.000357 0.002302 0.002532 0.001724 ...
#>  $ X4.02499962   : num  0.003417 0.000364 0.002225 0.002955 0.001875 ...
#>  $ X4.01499987   : num  0.002721 0.000293 0.002024 0.002641 0.00155 ...
#>  $ X4.00499964   : num  0.002934 0.000441 0.002968 0.003111 0.001966 ...
#>  $ X3.99499965   : num  0.00434 0.00137 0.00744 0.0067 0.0066 ...
#>  $ X3.98499966   : num  0.00566 0.0024 0.0119 0.01158 0.01207 ...
#>  $ X3.97499967   : num  0.00707 0.00324 0.01267 0.01634 0.01684 ...
#>  $ X3.96499968   : num  0.00563 0.00235 0.00826 0.01258 0.0123 ...
#>  $ X3.95499969   : num  0.00376 0.00117 0.00412 0.00751 0.00702 ...
#>  $ X3.94499969   : num  0.002652 0.000628 0.004445 0.00498 0.004357 ...
#>  $ X3.9349997    : num  0.00655 0.00281 0.02079 0.01522 0.01211 ...
#>  $ X3.92499971   : num  0.00363 0.00123 0.00905 0.00718 0.00681 ...
#>  $ X3.91499972   : num  0.002792 0.000865 0.004798 0.006672 0.006196 ...
#>  $ X3.90499973   : num  0.002578 0.000781 0.003784 0.004664 0.004461 ...
#>  $ X3.89499974   : num  0.002639 0.000759 0.003503 0.005601 0.005141 ...
#>  $ X3.88499975   : num  0.002802 0.000867 0.003552 0.005681 0.005107 ...
#>  $ X3.87499976   : num  0.002679 0.000654 0.003415 0.004034 0.003941 ...
#>  $ X3.86499977   : num  0.002779 0.000804 0.003319 0.003711 0.003696 ...
#>  $ X3.85499978   : num  0.00341 0.00178 0.00348 0.00454 0.00436 ...
#>  $ X3.84499979   : num  0.00402 0.00207 0.00397 0.0048 0.00475 ...
#>  $ X3.8349998    : num  0.00329 0.00104 0.00372 0.00594 0.00538 ...
#>  $ X3.82499981   : num  0.002469 0.000598 0.002882 0.005108 0.003755 ...
#>  $ X3.81499982   : num  0.001844 0.000403 0.002334 0.003874 0.002338 ...
#>  $ X3.80499983   : num  0.001687 0.000349 0.002409 0.002988 0.001682 ...
#>  $ X3.7949996    : num  0.002887 0.000936 0.004881 0.005173 0.004458 ...
#>  $ X3.78499961   : num  0.0058 0.00254 0.01103 0.01143 0.01135 ...
#>  $ X3.77499962   : num  0.00786 0.00373 0.01485 0.01407 0.01538 ...
#>  $ X3.76499963   : num  0.00625 0.00324 0.01118 0.00987 0.01087 ...
#>  $ X3.75499964   : num  0.00663 0.00408 0.00789 0.00782 0.00823 ...
#>  $ X3.74499965   : num  0.0099 0.00765 0.00661 0.00757 0.00722 ...
#>  $ X3.73499966   : num  0.01079 0.00887 0.0045 0.00583 0.00477 ...
#>  $ X3.72499967   : num  0.01506 0.01546 0.00503 0.00496 0.00387 ...
#>   [list output truncated]