A plsRbetamodel model on a data subset A precomputed four components plsRbetamodel fitted to a subset of an example dataset and used in the vignette.

A plsRbetamodel model on a data subset

A precomputed four components plsRbetamodel fitted to a subset of an example dataset and used in the vignette.

modpls_sub4

Format

a class plsRbetamodel object

References

Régression Bêta PLS. (French) [PLS Beta regression.], F. Bertrand, N. Meyer, M. Beau-Faller, K. El Bayed, N. Izzie-J., M. Maumy-Bertrand, (2013), J. SFdS, 154(3):143-159

Partial Least Squares Regression for Beta Regression Models. F. Bertrand, M. Maumy (2021). useR! 2021, Zurich.

Examples

data(modpls_sub4)
str(modpls_sub4)

#> List of 53
#>  $ nr              : int 80
#>  $ nc              : int 79
#>  $ nt              : num 10
#>  $ ww              : num [1:79, 1:10] -0.9319 -0.6852 -1.0027 -0.0806 0.0401 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:79] "X4.65499973" "X4.6449995" "X4.63499975" "X4.60499954" ...
#>   .. ..$ : chr [1:10] "Comp_ 1" "Comp_ 2" "Comp_ 3" "Comp_ 4" ...
#>  $ wwnorm          : num [1:79, 1:10] -0.20223 -0.1487 -0.2176 -0.0175 0.00871 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:79] "X4.65499973" "X4.6449995" "X4.63499975" "X4.60499954" ...
#>   .. ..$ : chr [1:10] "Comp_ 1" "Comp_ 2" "Comp_ 3" "Comp_ 4" ...
#>  $ wwetoile        : num [1:79, 1:10] -0.20223 -0.1487 -0.2176 -0.0175 0.00871 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:79] "X4.65499973" "X4.6449995" "X4.63499975" "X4.60499954" ...
#>   .. ..$ : chr [1:10] "Coord_Comp_ 1" "Coord_Comp_ 2" "Coord_Comp_ 3" "Coord_Comp_ 4" ...
#>  $ tt              : num [1:80, 1:10] -4.05 -2.1 -3.54 -9.02 -7.14 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:80] "    10lamcla-C_4_1" "    11mehjac-C_4_1" "    12hesson-C_4_1" "    13pormar-C_4_1" ...
#>   .. ..$ : chr [1:10] "Comp_ 1" "Comp_ 2" "Comp_ 3" "Comp_ 4" ...
#>  $ pp              : num [1:79, 1:10] -0.1988 -0.1672 -0.2 -0.0252 0.0139 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:79] "X4.65499973" "X4.6449995" "X4.63499975" "X4.60499954" ...
#>   .. ..$ : chr [1:10] "Comp_ 1" "Comp_ 2" "Comp_ 3" "Comp_ 4" ...
#>  $ CoeffC          : Named num [1:10] 0.416 0.128 0.124 0.26 0.2 ...
#>   ..- attr(*, "names")= chr [1:10] "Coeff_Comp_Reg 1" "Coeff_Comp_Reg 2" "Coeff_Comp_Reg 3" "Coeff_Comp_Reg 4" ...
#>  $ uscores         : NULL
#>  $ YChapeau        : num [1:80, 1] 0.071 0.082 0.0471 0.0167 0.0275 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:80] "    10lamcla-C_4_1" "    11mehjac-C_4_1" "    12hesson-C_4_1" "    13pormar-C_4_1" ...
#>   .. ..$ : NULL
#>  $ residYChapeau   : Named num [1:80] -2.57 -2.42 -3.01 -4.08 -3.57 ...
#>   ..- attr(*, "names")= chr [1:80] "    10lamcla-C_4_1" "    11mehjac-C_4_1" "    12hesson-C_4_1" "    13pormar-C_4_1" ...
#>  $ RepY            : Named num [1:80] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 ...
#>   ..- attr(*, "names")= chr [1:80] "    10lamcla-C_4_1" "    11mehjac-C_4_1" "    12hesson-C_4_1" "    13pormar-C_4_1" ...
#>   ..- attr(*, "scaled:center")= num 0
#>   ..- attr(*, "scaled:scale")= num 1
#>  $ na.miss.Y       : logi FALSE
#>  $ YNA             : Named logi [1:80] TRUE TRUE TRUE TRUE TRUE TRUE ...
#>   ..- attr(*, "names")= chr [1:80] "    10lamcla-C_4_1" "    11mehjac-C_4_1" "    12hesson-C_4_1" "    13pormar-C_4_1" ...
#>  $ residY          : Named num [1:80] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 ...
#>   ..- attr(*, "names")= chr [1:80] "    10lamcla-C_4_1" "    11mehjac-C_4_1" "    12hesson-C_4_1" "    13pormar-C_4_1" ...
#>   ..- attr(*, "scaled:center")= num 0
#>   ..- attr(*, "scaled:scale")= num 1
#>  $ ExpliX          : num [1:80, 1:79] 0.0454 -0.6218 -0.1947 3.2951 2.0219 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:80] "    10lamcla-C_4_1" "    11mehjac-C_4_1" "    12hesson-C_4_1" "    13pormar-C_4_1" ...
#>   .. ..$ : chr [1:79] "X4.65499973" "X4.6449995" "X4.63499975" "X4.60499954" ...
#>   ..- attr(*, "scaled:center")= Named num [1:79] 0.000811 0.000523 0.000607 0.000473 0.001042 ...
#>   .. ..- attr(*, "names")= chr [1:79] "X4.65499973" "X4.6449995" "X4.63499975" "X4.60499954" ...
#>   ..- attr(*, "scaled:scale")= Named num [1:79] 0.001031 0.000734 0.000759 0.000346 0.00043 ...
#>   .. ..- attr(*, "names")= chr [1:79] "X4.65499973" "X4.6449995" "X4.63499975" "X4.60499954" ...
#>  $ na.miss.X       : logi FALSE
#>  $ XXNA            : logi [1:80, 1:79] TRUE TRUE TRUE TRUE TRUE TRUE ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:80] "    10lamcla-C_4_1" "    11mehjac-C_4_1" "    12hesson-C_4_1" "    13pormar-C_4_1" ...
#>   .. ..$ : chr [1:79] "X4.65499973" "X4.6449995" "X4.63499975" "X4.60499954" ...
#>  $ residXX         : num [1:80, 1:79] -0.0239 0.1541 -0.1293 0.3929 0.4318 ...
#>   ..- attr(*, "scaled:center")= Named num [1:79] 0.000811 0.000523 0.000607 0.000473 0.001042 ...
#>   .. ..- attr(*, "names")= chr [1:79] "X4.65499973" "X4.6449995" "X4.63499975" "X4.60499954" ...
#>   ..- attr(*, "scaled:scale")= Named num [1:79] 0.001031 0.000734 0.000759 0.000346 0.00043 ...
#>   .. ..- attr(*, "names")= chr [1:79] "X4.65499973" "X4.6449995" "X4.63499975" "X4.60499954" ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:80] "    10lamcla-C_4_1" "    11mehjac-C_4_1" "    12hesson-C_4_1" "    13pormar-C_4_1" ...
#>   .. ..$ : chr [1:79] "X4.65499973" "X4.6449995" "X4.63499975" "X4.60499954" ...
#>  $ PredictY        : num [1:80, 1:79] 0.0454 -0.6218 -0.1947 3.2951 2.0219 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:80] "    10lamcla-C_4_1" "    11mehjac-C_4_1" "    12hesson-C_4_1" "    13pormar-C_4_1" ...
#>   .. ..$ : chr [1:79] "X4.65499973" "X4.6449995" "X4.63499975" "X4.60499954" ...
#>   ..- attr(*, "scaled:center")= Named num [1:79] 0.000811 0.000523 0.000607 0.000473 0.001042 ...
#>   .. ..- attr(*, "names")= chr [1:79] "X4.65499973" "X4.6449995" "X4.63499975" "X4.60499954" ...
#>   ..- attr(*, "scaled:scale")= Named num [1:79] 0.001031 0.000734 0.000759 0.000346 0.00043 ...
#>   .. ..- attr(*, "names")= chr [1:79] "X4.65499973" "X4.6449995" "X4.63499975" "X4.60499954" ...
#>  $ RSS             : num [1, 1:11] 6.44 2.29 1.97 1.88 1.53 ...
#>  $ RSSresidY       : num [1, 1:11] 6.44 189.03 204 209.1 227.18 ...
#>  $ R2              : num [1:10] 0.644 0.694 0.709 0.762 0.797 ...
#>  $ R2residY        : num [1:10] -28.3 -30.7 -31.4 -34.2 -36.9 ...
#>  $ press.ind       : NULL
#>  $ press.tot       : NULL
#>  $ Q2cum           : logi [1:10] NA NA NA NA NA NA ...
#>  $ family          : NULL
#>  $ ttPredictY      : num [1:80, 1:10] -4.05 -2.1 -3.54 -9.02 -7.14 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:80] "    10lamcla-C_4_1" "    11mehjac-C_4_1" "    12hesson-C_4_1" "    13pormar-C_4_1" ...
#>   .. ..$ : chr [1:10] "tt1" "tt2" "tt3" "tt4" ...
#>  $ typeVC          : chr "none"
#>  $ dataX           :'data.frame':	80 obs. of  79 variables:
#>   ..$ X4.65499973: num [1:80] 0.000858 0.00017 0.00061 0.00421 0.002897 ...
#>   ..$ X4.6449995 : num [1:80] 3.87e-04 3.59e-05 2.61e-04 3.05e-03 1.55e-03 ...
#>   ..$ X4.63499975: num [1:80] 0.000596 0.000147 0.000641 0.002869 0.001939 ...
#>   ..$ X4.60499954: num [1:80] 2.72e-04 1.92e-05 4.53e-04 8.31e-04 2.37e-04 ...
#>   ..$ X4.59499979: num [1:80] 4.90e-04 9.56e-05 9.26e-04 1.30e-03 8.26e-04 ...
#>   ..$ X4.58499956: num [1:80] 0.00092 0.000286 0.001638 0.001838 0.001668 ...
#>   ..$ X4.57499981: num [1:80] 0.001012 0.000348 0.001851 0.002019 0.001903 ...
#>   ..$ X4.56499958: num [1:80] 0.000713 0.000218 0.001074 0.001364 0.001047 ...
#>   ..$ X4.55499983: num [1:80] 3.83e-04 4.46e-05 3.68e-04 7.45e-04 2.24e-04 ...
#>   ..$ X4.52499962: num [1:80] 0.000966 0.000245 0.001627 0.00247 0.001954 ...
#>   ..$ X4.51499987: num [1:80] 0.001216 0.000362 0.001684 0.001861 0.001258 ...
#>   ..$ X4.50499964: num [1:80] 3.02e-04 2.53e-05 3.19e-04 7.99e-04 1.61e-04 ...
#>   ..$ X4.43499947: num [1:80] 0.000661 0.000124 0.000612 0.000984 0.000617 ...
#>   ..$ X4.42499971: num [1:80] 0.000582 0.00013 0.000809 0.001099 0.000516 ...
#>   ..$ X4.38499975: num [1:80] 0.000615 0.000165 0.001093 0.001322 0.000648 ...
#>   ..$ X4.37499952: num [1:80] 0.000751 0.000234 0.001365 0.00197 0.000956 ...
#>   ..$ X4.35499954: num [1:80] 0.000808 0.000213 0.001216 0.001665 0.000708 ...
#>   ..$ X4.2949996 : num [1:80] 0.00101 0.00039 0.00173 0.00208 0.00141 ...
#>   ..$ X4.28499985: num [1:80] 0.0011 0.00039 0.00185 0.00213 0.00143 ...
#>   ..$ X4.27499962: num [1:80] 0.001184 0.000436 0.002059 0.002134 0.001581 ...
#>   ..$ X4.26499987: num [1:80] 0.001126 0.000454 0.002035 0.002173 0.001327 ...
#>   ..$ X4.25499964: num [1:80] 0.001059 0.000412 0.001895 0.001904 0.001288 ...
#>   ..$ X4.20499992: num [1:80] 0.000693 0.000111 0.000945 0.001135 0.000223 ...
#>   ..$ X4.19499969: num [1:80] 0.000804 0.000172 0.00164 0.001528 0.000459 ...
#>   ..$ X4.18499994: num [1:80] 0.001246 0.000522 0.003401 0.002819 0.001436 ...
#>   ..$ X4.1449995 : num [1:80] 0.00371 0.00147 0.00797 0.00822 0.00499 ...
#>   ..$ X4.13499975: num [1:80] 0.00573 0.00211 0.01354 0.01176 0.00862 ...
#>   ..$ X4.12499952: num [1:80] 0.00593 0.00268 0.01487 0.01133 0.00941 ...
#>   ..$ X4.11499977: num [1:80] 0.00808 0.0038 0.01942 0.01456 0.01432 ...
#>   ..$ X4.10499954: num [1:80] 0.00405 0.00158 0.00821 0.00709 0.00651 ...
#>   ..$ X4.09499979: num [1:80] 0.002255 0.000585 0.003045 0.002778 0.002261 ...
#>   ..$ X4.05499983: num [1:80] 0.01216 0.00299 0.01732 0.01715 0.01925 ...
#>   ..$ X4.0449996 : num [1:80] 0.00717 0.00108 0.00794 0.00597 0.0066 ...
#>   ..$ X4.00499964: num [1:80] 0.002934 0.000441 0.002968 0.003111 0.001966 ...
#>   ..$ X3.99499965: num [1:80] 0.00434 0.00137 0.00744 0.0067 0.0066 ...
#>   ..$ X3.98499966: num [1:80] 0.00566 0.0024 0.0119 0.01158 0.01207 ...
#>   ..$ X3.85499978: num [1:80] 0.00341 0.00178 0.00348 0.00454 0.00436 ...
#>   ..$ X3.84499979: num [1:80] 0.00402 0.00207 0.00397 0.0048 0.00475 ...
#>   ..$ X3.8349998 : num [1:80] 0.00329 0.00104 0.00372 0.00594 0.00538 ...
#>   ..$ X3.82499981: num [1:80] 0.002469 0.000598 0.002882 0.005108 0.003755 ...
#>   ..$ X3.81499982: num [1:80] 0.001844 0.000403 0.002334 0.003874 0.002338 ...
#>   ..$ X3.80499983: num [1:80] 0.001687 0.000349 0.002409 0.002988 0.001682 ...
#>   ..$ X3.7949996 : num [1:80] 0.002887 0.000936 0.004881 0.005173 0.004458 ...
#>   ..$ X3.76499963: num [1:80] 0.00625 0.00324 0.01118 0.00987 0.01087 ...
#>   ..$ X3.75499964: num [1:80] 0.00663 0.00408 0.00789 0.00782 0.00823 ...
#>   ..$ X3.73499966: num [1:80] 0.01079 0.00887 0.0045 0.00583 0.00477 ...
#>   ..$ X3.70499969: num [1:80] 0.23211 0.41724 0.01811 0.00843 0.00662 ...
#>   ..$ X3.69499969: num [1:80] 0.01763 0.02793 0.00717 0.00714 0.00748 ...
#>   ..$ X3.66499972: num [1:80] 0.00469 0.00428 0.01194 0.00921 0.00916 ...
#>   ..$ X3.65499973: num [1:80] 0.0065 0.0068 0.00848 0.00763 0.00691 ...
#>   ..$ X3.61499977: num [1:80] 0.01357 0.00565 0.02878 0.02325 0.02749 ...
#>   ..$ X3.59499979: num [1:80] 0.01005 0.00423 0.01942 0.01516 0.01883 ...
#>   ..$ X3.56499982: num [1:80] 0.00274 0.00286 0.00339 0.00438 0.00328 ...
#>   ..$ X3.54499984: num [1:80] 0.01196 0.00461 0.02309 0.02705 0.0302 ...
#>   ..$ X3.53499961: num [1:80] 0.00728 0.00289 0.01541 0.01286 0.01348 ...
#>   ..$ X3.52499962: num [1:80] 0.00841 0.00348 0.01591 0.01911 0.02145 ...
#>   ..$ X3.51499963: num [1:80] 0.00449 0.00202 0.00932 0.00708 0.00807 ...
#>   ..$ X3.50499964: num [1:80] 0.000882 0.000718 0.001341 0.002757 0.002684 ...
#>   ..$ X3.49499965: num [1:80] 0.000533 0.000497 0.000572 0.001676 0.001368 ...
#>   ..$ X3.48499966: num [1:80] 0.00095 0.00069 0.00109 0.00489 0.00393 ...
#>   ..$ X3.47499967: num [1:80] 0.000624 0.000537 0.000793 0.003055 0.002235 ...
#>   ..$ X3.46499968: num [1:80] 0.001016 0.000655 0.001267 0.005273 0.004053 ...
#>   ..$ X3.44499969: num [1:80] 0.001334 0.000657 0.001646 0.004326 0.003092 ...
#>   ..$ X3.4349997 : num [1:80] 0.00258 0.00128 0.00505 0.00461 0.00488 ...
#>   ..$ X3.42499971: num [1:80] 0.00389 0.00183 0.00753 0.00774 0.00878 ...
#>   ..$ X3.41499972: num [1:80] 0.0038 0.00164 0.00617 0.00786 0.00718 ...
#>   ..$ X3.40499973: num [1:80] 0.00361 0.0016 0.00608 0.00675 0.00665 ...
#>   ..$ X3.37499976: num [1:80] 0.000497 0.000363 0.000566 0.001572 0.001171 ...
#>   ..$ X3.36499977: num [1:80] 0.000574 0.000357 0.00073 0.000884 0.000574 ...
#>   ..$ X3.35499978: num [1:80] 0.003036 0.000476 0.001356 0.001738 0.001889 ...
#>   ..$ X3.32499981: num [1:80] 0.000474 0.000354 0.000936 0.001009 0.001228 ...
#>   ..$ X3.31499982: num [1:80] 0.00043 0.000312 0.000822 0.000895 0.000539 ...
#>   ..$ X3.30499983: num [1:80] 0.000486 0.000311 0.000837 0.001323 0.000815 ...
#>   ..$ X3.27499962: num [1:80] 0.00824 0.0033 0.01527 0.01752 0.01908 ...
#>   ..$ X3.26499963: num [1:80] 0.00761 0.0027 0.01755 0.0144 0.01577 ...
#>   ..$ X3.25499964: num [1:80] 0.00716 0.00272 0.01312 0.0132 0.01324 ...
#>   ..$ X3.24499965: num [1:80] 0.00422 0.00168 0.00823 0.01007 0.00808 ...
#>   ..$ X3.19499969: num [1:80] 0.00357 0.00209 0.00608 0.00602 0.00956 ...
#>   ..$ X3.13499975: num [1:80] 0.000637 0.000466 0.000795 0.001664 0.001634 ...
#>  $ dataY           : Named num [1:80] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 ...
#>   ..- attr(*, "names")= chr [1:80] "    10lamcla-C_4_1" "    11mehjac-C_4_1" "    12hesson-C_4_1" "    13pormar-C_4_1" ...
#>  $ weights         : int [1:80] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ computed_nt     : int 10
#>  $ AIC             : num [1, 1:11] -32.3 -129.6 -134.6 -142.7 -155.5 ...
#>  $ BIC             : num [1, 1:11] -27.6 -122.4 -125.1 -130.8 -141.2 ...
#>  $ Coeffsmodel_vals: num [1:11, 1:44] -0.637 NA NA NA NA ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:11] "(Intercept)" "" "" "" ...
#>   .. ..$ : chr [1:44] "Estimate" "Std. Error" "z value" "Pr(>|z|)" ...
#>  $ ChisqPearson    : num [1:11] 76.8 79.5 72.2 79 84.1 ...
#>  $ pseudo.R2       : num [1, 1:10] 0.776 0.799 0.828 0.856 0.874 ...
#>  $ CoeffCFull      : num [1:11, 1:10] -0.7 0.325 NA NA NA ...
#>  $ CoeffConstante  : num [1, 1:10] -0.7 -0.726 -0.723 -0.751 -0.773 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:10] "" "tempCoeffConstante" "tempCoeffConstante" "tempCoeffConstante" ...
#>  $ Std.Coeffs      : num [1:80, 1] -0.9051 0.0631 0.2246 -0.2159 0.0452 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:80] "Intercept" "X4.65499973" "X4.6449995" "X4.63499975" ...
#>   .. ..$ : NULL
#>  $ Coeffs          : num [1:80, 1] 1.54 61.18 306.05 -284.41 130.57 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:80] "Intercept" "X4.65499973" "X4.6449995" "X4.63499975" ...
#>   .. ..$ : NULL
#>  $ Yresidus        : Named num [1:80] -0.06096 -0.07196 -0.03708 -0.00666 -0.0175 ...
#>   ..- attr(*, "names")= chr [1:80] "    10lamcla-C_4_1" "    11mehjac-C_4_1" "    12hesson-C_4_1" "    13pormar-C_4_1" ...
#>  $ residusY        : num [1:80, 1:10] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:80] "    10lamcla-C_4_1" "    11mehjac-C_4_1" "    12hesson-C_4_1" "    13pormar-C_4_1" ...
#>   .. ..$ : NULL
#>  $ InfCrit         : num [1:11, 1:6] -32.3 -129.6 -134.6 -142.7 -155.5 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:11] "Nb_Comp_0" "Nb_Comp_1" "Nb_Comp_2" "Nb_Comp_3" ...
#>   .. ..$ : chr [1:6] "AIC" "BIC" "Chi2_Pearson_Y" "RSS_Y" ...
#>  $ Std.ValsPredictY: Named num [1:80] 0.071 0.082 0.0471 0.0167 0.0275 ...
#>   ..- attr(*, "names")= chr [1:80] "    10lamcla-C_4_1" "    11mehjac-C_4_1" "    12hesson-C_4_1" "    13pormar-C_4_1" ...
#>  $ ValsPredictY    : Named num [1:80] 0.071 0.082 0.0471 0.0167 0.0275 ...
#>   ..- attr(*, "names")= chr [1:80] "    10lamcla-C_4_1" "    11mehjac-C_4_1" "    12hesson-C_4_1" "    13pormar-C_4_1" ...
#>  $ Std.XChapeau    : num [1:80, 1:79] 0.0692 -0.7759 -0.0654 2.9022 1.5901 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:80] "    10lamcla-C_4_1" "    11mehjac-C_4_1" "    12hesson-C_4_1" "    13pormar-C_4_1" ...
#>   .. ..$ : NULL
#>  $ FinalModel      :List of 29
#>   ..$ coefficients    :List of 2
#>   .. ..$ mean     : Named num [1:11] -0.905 0.416 0.128 0.124 0.26 ...
#>   .. .. ..- attr(*, "names")= chr [1:11] "(Intercept)" "tt1" "tt2" "tt3" ...
#>   .. ..$ precision: Named num 19.1
#>   .. .. ..- attr(*, "names")= chr "(phi)"
#>   ..$ residuals       : Named num [1:80] -0.06096 -0.07196 -0.03708 -0.00666 -0.0175 ...
#>   .. ..- attr(*, "names")= chr [1:80] "    10lamcla-C_4_1" "    11mehjac-C_4_1" "    12hesson-C_4_1" "    13pormar-C_4_1" ...
#>   ..$ fitted.values   : Named num [1:80] 0.071 0.082 0.0471 0.0167 0.0275 ...
#>   .. ..- attr(*, "names")= chr [1:80] "    10lamcla-C_4_1" "    11mehjac-C_4_1" "    12hesson-C_4_1" "    13pormar-C_4_1" ...
#>   ..$ type            : chr "ML"
#>   ..$ optim           :List of 6
#>   .. ..$ par        : Named num [1:12] -0.905 0.416 0.128 0.124 0.26 ...
#>   .. .. ..- attr(*, "names")= chr [1:12] "mean.(Intercept)" "mean.tt1" "mean.tt2" "mean.tt3" ...
#>   .. ..$ value      : num 107
#>   .. ..$ counts     : Named int [1:2] 150 44
#>   .. .. ..- attr(*, "names")= chr [1:2] "function" "gradient"
#>   .. ..$ convergence: int 0
#>   .. ..$ message    : NULL
#>   .. ..$ hessian    : num [1:12, 1:12] -290.41 -405.09 -67.27 -21.01 6.84 ...
#>   .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. ..$ : chr [1:12] "mean.(Intercept)" "mean.tt1" "mean.tt2" "mean.tt3" ...
#>   .. .. .. ..$ : chr [1:12] "mean.(Intercept)" "mean.tt1" "mean.tt2" "mean.tt3" ...
#>   ..$ method          : chr "BFGS"
#>   ..$ control         :List of 10
#>   .. ..$ phi    : logi FALSE
#>   .. ..$ method : chr "BFGS"
#>   .. ..$ maxit  : num 5000
#>   .. ..$ hessian: logi TRUE
#>   .. ..$ trace  : logi FALSE
#>   .. ..$ start  : NULL
#>   .. ..$ fsmaxit: num 200
#>   .. ..$ fstol  : num 1e-08
#>   .. ..$ fnscale: num -1
#>   .. ..$ reltol : num 9.02e-14
#>   ..$ scoring         : num 0
#>   ..$ start           : Named num [1:12] -1.206 0.507 0.129 0.185 0.315 ...
#>   .. ..- attr(*, "names")= chr [1:12] "mean.(Intercept)" "mean.tt1" "mean.tt2" "mean.tt3" ...
#>   ..$ weights         : NULL
#>   ..$ offset          :List of 2
#>   .. ..$ mean     : NULL
#>   .. ..$ precision: NULL
#>   ..$ n               : int 80
#>   ..$ nobs            : int 80
#>   ..$ df.null         : num 78
#>   ..$ df.residual     : int 68
#>   ..$ phi             : logi FALSE
#>   ..$ loglik          : num 107
#>   ..$ vcov            : num [1:12, 1:12] 5.18e-03 -9.78e-04 -3.25e-04 2.41e-06 -3.17e-05 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : chr [1:12] "(Intercept)" "tt1" "tt2" "tt3" ...
#>   .. .. ..$ : chr [1:12] "(Intercept)" "tt1" "tt2" "tt3" ...
#>   ..$ bias            : Named num [1:12] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..- attr(*, "names")= chr [1:12] "(Intercept)" "tt1" "tt2" "tt3" ...
#>   ..$ pseudo.r.squared: num 0.906
#>   ..$ link            :List of 2
#>   .. ..$ mean     :List of 6
#>   .. .. ..$ linkfun :function (mu)  
#>   .. .. ..$ linkinv :function (eta)  
#>   .. .. ..$ mu.eta  :function (eta)  
#>   .. .. ..$ valideta:function (eta)  
#>   .. .. ..$ name    : chr "logit"
#>   .. .. ..$ dmu.deta:function (eta)  
#>   .. .. ..- attr(*, "class")= chr "link-glm"
#>   .. ..$ precision:List of 6
#>   .. .. ..$ linkfun :function (mu)  
#>   .. .. ..$ linkinv :function (eta)  
#>   .. .. ..$ mu.eta  :function (eta)  
#>   .. .. ..$ valideta:function (eta)  
#>   .. .. ..$ name    : chr "identity"
#>   .. .. ..$ dmu.deta:function (eta)  
#>   .. .. ..- attr(*, "class")= chr "link-glm"
#>   ..$ converged       : logi TRUE
#>   ..$ call            : language betareg(formula = YwotNA ~ tt, na.action = function (object, ...)  UseMethod("na.exclude"), link = "logit", type | __truncated__ ...
#>   ..$ formula         :Class 'formula'  language YwotNA ~ tt
#>   .. .. ..- attr(*, ".Environment")=<environment: 0x7fa696ffd5b0> 
#>   ..$ terms           :List of 3
#>   .. ..$ mean     :Classes 'terms', 'formula'  language YwotNA ~ tt
#>   .. .. .. ..- attr(*, "variables")= language list(YwotNA, tt)
#>   .. .. .. ..- attr(*, "factors")= int [1:2, 1] 0 1
#>   .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. ..$ : chr [1:2] "YwotNA" "tt"
#>   .. .. .. .. .. ..$ : chr "tt"
#>   .. .. .. ..- attr(*, "term.labels")= chr "tt"
#>   .. .. .. ..- attr(*, "order")= int 1
#>   .. .. .. ..- attr(*, "intercept")= int 1
#>   .. .. .. ..- attr(*, "response")= int 1
#>   .. .. .. ..- attr(*, ".Environment")=<environment: 0x7fa696ffd5b0> 
#>   .. ..$ precision:Classes 'terms', 'formula'  language ~1
#>   .. .. .. ..- attr(*, "variables")= language list()
#>   .. .. .. ..- attr(*, "factors")= int(0) 
#>   .. .. .. ..- attr(*, "term.labels")= chr(0) 
#>   .. .. .. ..- attr(*, "order")= int(0) 
#>   .. .. .. ..- attr(*, "intercept")= int 1
#>   .. .. .. ..- attr(*, "response")= num 0
#>   .. .. .. ..- attr(*, ".Environment")=<environment: 0x7fa696ffd5b0> 
#>   .. ..$ full     :Classes 'terms', 'formula'  language YwotNA ~ tt + 1
#>   .. .. .. ..- attr(*, "variables")= language list(YwotNA, tt)
#>   .. .. .. ..- attr(*, "factors")= int [1:2, 1] 0 1
#>   .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. ..$ : chr [1:2] "YwotNA" "tt"
#>   .. .. .. .. .. ..$ : chr "tt"
#>   .. .. .. ..- attr(*, "term.labels")= chr "tt"
#>   .. .. .. ..- attr(*, "order")= int 1
#>   .. .. .. ..- attr(*, "intercept")= int 1
#>   .. .. .. ..- attr(*, "response")= int 1
#>   .. .. .. ..- attr(*, ".Environment")=<environment: 0x7fa696ffd5b0> 
#>   ..$ levels          :List of 3
#>   .. ..$ mean     : Named list()
#>   .. ..$ precision: NULL
#>   .. ..$ full     : Named list()
#>   ..$ contrasts       :List of 2
#>   .. ..$ mean     : NULL
#>   .. ..$ precision: NULL
#>   ..$ model           :'data.frame':	80 obs. of  2 variables:
#>   .. ..$ YwotNA: num [1:80, 1] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 ...
#>   .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. ..$ : chr [1:80] "    10lamcla-C_4_1" "    11mehjac-C_4_1" "    12hesson-C_4_1" "    13pormar-C_4_1" ...
#>   .. .. .. ..$ : NULL
#>   .. ..$ tt    : num [1:80, 1:10] -4.05 -2.1 -3.54 -9.02 -7.14 ...
#>   .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. ..$ : chr [1:80] "    10lamcla-C_4_1" "    11mehjac-C_4_1" "    12hesson-C_4_1" "    13pormar-C_4_1" ...
#>   .. .. .. ..$ : NULL
#>   .. ..- attr(*, "terms")=Classes 'terms', 'formula'  language YwotNA ~ tt + 1
#>   .. .. .. ..- attr(*, "variables")= language list(YwotNA, tt)
#>   .. .. .. ..- attr(*, "factors")= int [1:2, 1] 0 1
#>   .. .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. .. ..$ : chr [1:2] "YwotNA" "tt"
#>   .. .. .. .. .. ..$ : chr "tt"
#>   .. .. .. ..- attr(*, "term.labels")= chr "tt"
#>   .. .. .. ..- attr(*, "order")= int 1
#>   .. .. .. ..- attr(*, "intercept")= int 1
#>   .. .. .. ..- attr(*, "response")= int 1
#>   .. .. .. ..- attr(*, ".Environment")=<environment: 0x7fa696ffd5b0> 
#>   .. .. .. ..- attr(*, "predvars")= language list(YwotNA, tt)
#>   .. .. .. ..- attr(*, "dataClasses")= Named chr [1:2] "nmatrix.1" "nmatrix.10"
#>   .. .. .. .. ..- attr(*, "names")= chr [1:2] "YwotNA" "tt"
#>   ..$ y               : Named num [1:80] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 ...
#>   .. ..- attr(*, "names")= chr [1:80] "    10lamcla-C_4_1" "    11mehjac-C_4_1" "    12hesson-C_4_1" "    13pormar-C_4_1" ...
#>   ..- attr(*, "class")= chr "betareg"
#>  $ XXwotNA         : num [1:80, 1:79] 0.0228 0.3376 -0.2162 0.6475 0.3771 ...
#>   ..- attr(*, "scaled:center")= Named num [1:79] 0.000811 0.000523 0.000607 0.000473 0.001042 ...
#>   .. ..- attr(*, "names")= chr [1:79] "X4.65499973" "X4.6449995" "X4.63499975" "X4.60499954" ...
#>   ..- attr(*, "scaled:scale")= Named num [1:79] 0.001031 0.000734 0.000759 0.000346 0.00043 ...
#>   .. ..- attr(*, "names")= chr [1:79] "X4.65499973" "X4.6449995" "X4.63499975" "X4.60499954" ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:80] "    10lamcla-C_4_1" "    11mehjac-C_4_1" "    12hesson-C_4_1" "    13pormar-C_4_1" ...
#>   .. ..$ : chr [1:79] "X4.65499973" "X4.6449995" "X4.63499975" "X4.60499954" ...
#>  $ call            : language plsRbetamodel.formula(formula = X..Cellules.tumorales ~ ., data = colon_sub4,      nt = 10, modele = "pls-beta")
#>  - attr(*, "class")= chr "plsRbetamodel"