Coefficients function for permutation bootstrap techniques

A function passed to boot to perform bootstrap.

permcoefs.plsRbeta(
  dataset,
  ind,
  nt,
  modele,
  family = NULL,
  method = "logistic",
  link = "logit",
  link.phi = NULL,
  type = "ML",
  maxcoefvalues,
  ifbootfail,
  verbose = TRUE
)

Arguments

dataset

dataset to resample

ind

indices for resampling

nt

number of components to use

modele

name of the PLS glm or PLS beta model to be fitted ("pls", "pls-glm-Gamma", "pls-glm-gaussian", "pls-glm-inverse.gaussian", "pls-glm-logistic", "pls-glm-poisson", "pls-glm-polr", "pls-beta"). Use "modele=pls-glm-family" to enable the family option.

family

family to use if GLM model, see plsRbeta

method

method for beta regression

link

link for beta regression

link.phi

link.phi for beta regression

type

type of estimates

maxcoefvalues

maximum values allowed for the estimates of the coefficients to discard those coming from singular bootstrap samples

ifbootfail

value to return if the estimation fails on a

verbose

should info messages be displayed ?

Value

Estimates on a bootstrap sample.

References

Frédéric Bertrand, Nicolas Meyer, Michèle Beau-Faller, Karim El Bayed, Izzie-Jacques Namer, Myriam Maumy-Bertrand (2013). Régression Bêta PLS. Journal de la Société Française de Statistique, 154(3):143-159. http://publications-sfds.math.cnrs.fr/index.php/J-SFdS/article/view/215

See also

Author

Frédéric Bertrand
frederic.bertrand@utt.fr
https://fbertran.github.io/homepage/

Examples

# \donttest{
data("GasolineYield",package="betareg")

GazYield.boot <- bootplsbeta(plsRbeta(yield~.,data=GasolineYield,nt=3,
modele="pls-beta", verbose=FALSE), sim="ordinary", stype="i", R=250, statistic=permcoefs.plsRbeta)
# }